+++ testing.R
@@ -4,7 +4,7 @@ library(survival)
 KM <- survfit(Surv(rectime, censrec) ~ hormon, data = brcancer)
 time_scale <- seq(0, max(brcancer$rectime), by = 365)
 p <- KMunicate(fit = KM, time_scale = time_scale)
-pm1 <- KMunicate(fit = KM, time_scale = time_scale, .reverse = TRUE)
+KMunicate(fit = KM, time_scale = time_scale, .reverse = TRUE, .risk_table_base_size = 10, .rel_heights = c(5, 1, 1))
 
 ###
 devtools::check_win_devel(quiet = TRUE)