+++ NEWS.md
@@ -1,3 +1,5 @@
+# KMunicate (development version)
+
 # KMunicate 0.0.2
 
 * Initial release.