+++ NEWS.md
@@ -3,7 +3,8 @@
 ### New features
 
 * New argument `.reverse` to plot 1 - survival rather than survival;
-* New argument `.ylab` to customise the label for the vertical axis.
+* New argument `.ylab` to customise the label for the vertical axis;
+* New argument `.risk_table_base_size` to customise the base font size for the risk table.
 
 # KMunicate 0.0.2