+++ NEWS.md
@@ -4,7 +4,8 @@
 
 * New argument `.reverse` to plot 1 - survival rather than survival;
 * New argument `.ylab` to customise the label for the vertical axis;
-* New argument `.risk_table_base_size` to customise the base font size for the risk table.
+* New argument `.risk_table_base_size` to customise the base font size for the risk table;
+* Small edits to default plot style.
 
 # KMunicate 0.0.2