+++ NEWS.md
@@ -1,4 +1,4 @@
-# KMunicate (development version)
+# KMunicate 0.1.0
 
 ### New features